+++ XamlPowerToys.UI/Properties/AssemblyInfo.cs
@@ -3,11 +3,11 @@ using System.Runtime.InteropServices;
 using System.Windows;
 
 [assembly: AssemblyTitle("XamlPowerToys.UI")]
-[assembly: AssemblyDescription("Visual Studio 2015 - 2017 extension that enables developers to rapidly generate data entry forms from view models or entity objects for Xamarin Forms, UWP, WPF, and Silverlight.")]
+[assembly: AssemblyDescription("Visual Studio 2017 - 2019 extension that enables developers to rapidly generate data entry forms from view models or entity objects for Xamarin Forms, UWP, WPF, and Silverlight.")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
 [assembly: AssemblyProduct("XamlPowerToys.UI")]
-[assembly: AssemblyCopyright("Copyright © 2008-2017")]
+[assembly: AssemblyCopyright("Copyright © 2008-2019")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 [assembly: ComVisible(false)]
@@ -15,5 +15,5 @@ using System.Windows;
                ResourceDictionaryLocation.None,
                ResourceDictionaryLocation.SourceAssembly
            )]
-[assembly: AssemblyVersion("1.5.2.*")]
-[assembly: AssemblyFileVersion("1.5.2.0")]
+[assembly: AssemblyVersion("2.0.0.*")]
+[assembly: AssemblyFileVersion("2.0.0.0")]