+++ XamlPowerToys/PackageCommands/XamlPowerToysCommand.cs
@@ -1,62 +1,43 @@
 ﻿namespace XamlPowerToys.PackageCommands {
+
     using System;
     using System.ComponentModel.Design;
     using EnvDTE;
     using EnvDTE80;
-    using Microsoft.VisualStudio.CommandBars;
     using Microsoft.VisualStudio.Shell;
-    using Microsoft.VisualStudio.Shell.Interop;
     using XamlPowerToys.Commands;
-    using XamlPowerToys.Infrastructure;
-    using XamlPowerToys.Model;
+    using Task = System.Threading.Tasks.Task;
 
     internal sealed class XamlPowerToysCommand {
-
-        public DTE2 Dte2 { get; }
-
-        public static XamlPowerToysCommand Instance { get; private set; }
-
-        public Package Package { get; }
-
-        public IServiceProvider ServiceProvider => this.Package;
-
+        static DTE2 _dte2;
         public const Int32 CommandId = 0x0100;
-
         public static readonly Guid CommandSet = new Guid("4C87B692-1202-46AA-B64C-EF01FAEC53DA");
 
-        public static void Initialize(Package package) {
-            Instance = new XamlPowerToysCommand(package);
+        static void MenuItemCallback(Object sender, EventArgs e) {
+            ThreadHelper.ThrowIfNotOnUIThread();
+            if (_dte2.ActiveDocument.Name.EndsWith(".xaml")) {  // this is not needed but more insurance since I can't hide the menu item.
+                var dataFormGenerator = new DataFormGenerator(_dte2);
+                dataFormGenerator.Generate();
+            }
         }
 
-        XamlPowerToysCommand(Package package) {
-            if (package == null) {
-                throw new ArgumentNullException(nameof(package));
+        public static async Task InitializeAsync(AsyncPackage package) {
+            var dte = await package.GetServiceAsync(typeof(DTE));
+            if (dte == null) {
+                return;
             }
-            this.Package = package;
-            this.Dte2 = (DTE2)this.ServiceProvider.GetService(typeof(DTE));
+            _dte2 = (DTE2)dte;
 
-            var commandService = this.ServiceProvider.GetService(typeof(IMenuCommandService)) as OleMenuCommandService;
-            if (commandService != null) {
+            if (await package.GetServiceAsync(typeof(IMenuCommandService)) is IMenuCommandService commandService) {
                 var menuCommandID = new CommandID(CommandSet, CommandId);
-                var menuItem = new OleMenuCommand(MenuItemCallback, menuCommandID);
-                menuItem.BeforeQueryStatus += BeforeQueryStatusCallback;
+                var menuItem = new OleMenuCommand(MenuItemCallback, menuCommandID) {
+                    // this does not work, found here https://github.com/Microsoft/VSSDK-Extensibility-Samples/blob/master/SingleFileGenerator/src/Commands/ApplyCustomTool.cs
+                    // This will defer visibility control to the VisibilityConstraints section in the .vsct file
+                    Supported = false
+                };
+
                 commandService.AddCommand(menuItem);
             }
         }
-
-        void BeforeQueryStatusCallback(Object sender, EventArgs e) {
-            // gets called from a XAML file being edited by the WPF Designer editor.
-            var result = AssemblyAssistant.GetProjectType(this.Dte2.ActiveDocument.ActiveWindow.Project);
-            var cmd = (OleMenuCommand)sender;
-            // future proofing when Visual Studio changes the editor for Xamarin
-            cmd.Visible = result == ProjectType.Wpf || result == ProjectType.Xamarin || result == ProjectType.Uwp || result == ProjectType.Silverlight;
-
-        }
-
-        void MenuItemCallback(Object sender, EventArgs e) {
-            var dataFormGenerator = new DataFormGenerator(this.Dte2, this.Dte2.ActiveDocument.ActiveWindow.Project);
-            dataFormGenerator.Generate();
-        }
-
     }
 }